// src/api/models/Resource.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Resource = {
  readonly id: number;
  readonly link: string | null;
  readonly created: string;
  readonly modified: string;
  name: string;
  description?: string | null;
  url: string;
};